+++ src/app/inobeta-ui/ui/material-forms/controls/padding.ts
@@ -1,5 +1,6 @@
 import { Component, Input } from '@angular/core';
 import { IbFormControlBase, IbFormControlBaseComponent, IbFormControlBaseParams, IbFormControlData, IbFormControlInterface } from '../../forms/controls/form-control-base';
+/** @deprecated */
 @Component({
   selector: '[ib-mat-padding]',
   template: ``
@@ -7,7 +8,7 @@ import { IbFormControlBase, IbFormControlBaseComponent, IbFormControlBaseParams,
 export class IbMatPaddingComponent implements IbFormControlInterface {
   @Input() data: IbFormControlData;
 }
-
+/** @deprecated */
 export class IbMatPaddingControl extends IbFormControlBase<string>{
   constructor(options: IbFormControlBaseParams<string>){
     super(options)