+++ src/app/inobeta-ui/ui/material-forms/material-form-array/material-form-array.component.ts
@@ -1,6 +1,7 @@
 import { Component } from "@angular/core";
 import { IbDynamicFormArrayComponent } from "../../forms/dynamic-form-array/dynamic-form-array.component";
 
+/** @deprecated */
 @Component({
   selector: "ib-material-form-array",
   templateUrl: "material-form-array.component.html",