+++ src/app/inobeta-ui/ui/material-forms/material-form.stories.ts
@@ -17,7 +17,7 @@ import { IbMaterialFormModule } from "./material-form.module";
 import { IbMaterialFormComponent } from "./material-form/material-form.component";
 
 const meta: Meta = {
-  title: "Components/Forms",
+  title: "Components/Forms (deprecated)",
   component: IbMaterialFormComponent,
   decorators: [
     moduleMetadata({