+++ src/app/inobeta-ui/ui/modal/modal-message.component.ts
@@ -6,7 +6,12 @@ import { IbModalMessage } from './modal-message.model';
   selector: 'ib-modal-message',
   template: `
     <h2 mat-dialog-title>{{ data.title | translate }}</h2>
-    <mat-dialog-content style="min-width:350px;min-height: 10vh;">{{ data.message | translate }}</mat-dialog-content>
+    <mat-dialog-content
+      [ngStyle]="{
+        'min-width': this.data.minWidth ?? '350px',
+        'min-height': this.data.minHeight ?? '10vh'
+      }"
+    >{{ data.message | translate }}</mat-dialog-content>
     <mat-dialog-actions align="end">
       <button
         *ngFor="let btn of data.actions"