+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 | -- | -- | -- | -- |
 |DEVK-262|feat|Forms|Added `ib-form-control-errors` component and `[ibPatchFormValue]` directive|
 |DEVK-363|feat|all|Added support for Angular 18|
+|DEVK-405|fix|IbMaterialForms|Disabled fields ambiguity resolved|
 
 ### [18.0.1] - Breaking changes